+++ internal/encoding/ssh/filexfer/attrs.go
