+++ internal/encoding/ssh/filexfer/attrs_test.go
