+++ internal/encoding/ssh/filexfer/extensions_test.go
