+++ internal/encoding/ssh/filexfer/init_packets.go
