+++ internal/encoding/ssh/filexfer/open_packets.go
