+++ internal/encoding/ssh/filexfer/openssh/posix-rename.go
@@ -1,36 +1,36 @@
 package openssh
 
 import (
-	sshfx "github.com/pkg/sftp/encoding/ssh/filexfer"
+	sshfx "github.com/pkg/sftp/internal/encoding/ssh/filexfer"
 )
 
-const extensionHardlink = "hardlink@openssh.com"
+const extensionPosixRename = "posix-rename@openssh.com"
 
-// RegisterExtensionHardlink registers the "hardlink@openssh.com" extended packet with the encoding/ssh/filexfer package.
-func RegisterExtensionHardlink() {
-	sshfx.RegisterExtendedPacketType(extensionHardlink, func() sshfx.ExtendedData {
-		return new(HardlinkExtendedPacket)
+// RegisterExtensionPosixRename registers the "posix-rename@openssh.com" extended packet with the encoding/ssh/filexfer package.
+func RegisterExtensionPosixRename() {
+	sshfx.RegisterExtendedPacketType(extensionPosixRename, func() sshfx.ExtendedData {
+		return new(PosixRenameExtendedPacket)
 	})
 }
 
-// ExtensionHardlink returns an ExtensionPair suitable to append into an sshfx.InitPacket or sshfx.VersionPacket.
-func ExtensionHardlink() *sshfx.ExtensionPair {
+// ExtensionPosixRename returns an ExtensionPair suitable to append into an sshfx.InitPacket or sshfx.VersionPacket.
+func ExtensionPosixRename() *sshfx.ExtensionPair {
 	return &sshfx.ExtensionPair{
-		Name: extensionHardlink,
+		Name: extensionPosixRename,
 		Data: "1",
 	}
 }
 
-// HardlinkExtendedPacket defines the hardlink@openssh.com extend packet.
-type HardlinkExtendedPacket struct {
+// PosixRenameExtendedPacket defines the posix-rename@openssh.com extend packet.
+type PosixRenameExtendedPacket struct {
 	OldPath string
 	NewPath string
 }
 
 // MarshalPacket returns ep as a two-part binary encoding of the full extended packet.
-func (ep *HardlinkExtendedPacket) MarshalPacket(reqid uint32, b []byte) (header, payload []byte, err error) {
+func (ep *PosixRenameExtendedPacket) MarshalPacket(reqid uint32, b []byte) (header, payload []byte, err error) {
 	p := &sshfx.ExtendedPacket{
-		ExtendedRequest: extensionHardlink,
+		ExtendedRequest: extensionPosixRename,
 
 		Data: ep,
 	}
@@ -38,7 +38,7 @@ func (ep *HardlinkExtendedPacket) MarshalPacket(reqid uint32, b []byte) (header,
 }
 
 // MarshalInto encodes ep into the binary encoding of the hardlink@openssh.com extended packet-specific data.
-func (ep *HardlinkExtendedPacket) MarshalInto(buf *sshfx.Buffer) {
+func (ep *PosixRenameExtendedPacket) MarshalInto(buf *sshfx.Buffer) {
 	buf.AppendString(ep.OldPath)
 	buf.AppendString(ep.NewPath)
 }
@@ -46,7 +46,7 @@ func (ep *HardlinkExtendedPacket) MarshalInto(buf *sshfx.Buffer) {
 // MarshalBinary encodes ep into the binary encoding of the hardlink@openssh.com extended packet-specific data.
 //
 // NOTE: This _only_ encodes the packet-specific data, it does not encode the full extended packet.
-func (ep *HardlinkExtendedPacket) MarshalBinary() ([]byte, error) {
+func (ep *PosixRenameExtendedPacket) MarshalBinary() ([]byte, error) {
 	// string(oldpath) + string(newpath)
 	size := 4 + len(ep.OldPath) + 4 + len(ep.NewPath)
 
@@ -56,7 +56,7 @@ func (ep *HardlinkExtendedPacket) MarshalBinary() ([]byte, error) {
 }
 
 // UnmarshalFrom decodes the hardlink@openssh.com extended packet-specific data from buf.
-func (ep *HardlinkExtendedPacket) UnmarshalFrom(buf *sshfx.Buffer) (err error) {
+func (ep *PosixRenameExtendedPacket) UnmarshalFrom(buf *sshfx.Buffer) (err error) {
 	if ep.OldPath, err = buf.ConsumeString(); err != nil {
 		return err
 	}
@@ -69,6 +69,6 @@ func (ep *HardlinkExtendedPacket) UnmarshalFrom(buf *sshfx.Buffer) (err error) {
 }
 
 // UnmarshalBinary decodes the hardlink@openssh.com extended packet-specific data into ep.
-func (ep *HardlinkExtendedPacket) UnmarshalBinary(data []byte) (err error) {
+func (ep *PosixRenameExtendedPacket) UnmarshalBinary(data []byte) (err error) {
 	return ep.UnmarshalFrom(sshfx.NewBuffer(data))
 }
