+++ internal/encoding/ssh/filexfer/path_packets.go
