+++ internal/encoding/ssh/filexfer/path_packets_test.go
