+++ internal/encoding/ssh/filexfer/permissions.go
